requirements.txt
alembic==1.17.2
beautifulsoup4==4.14.3
blinker==1.9.0
click==8.3.1
docutils==0.22.4
et-xmlfile==1.1.0
flask==3.1.2
flask-compress>=1.23
flask-login==0.6.3
flask-mail==0.10.0
flask-migrate==4.1.0
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
greenlet==3.3.0
gunicorn==21.2.0
itsdangerous==2.2.0
jinja2==3.1.6
mako==1.3.10
markupsafe==2.1.3
numpy==1.26.2
openpyxl==3.1.5
packaging==23.2
pandas==2.3.3
playwright==1.57.0
psycopg2-binary>=2.9.11
pygments==2.19.2
pytest==8.1.1
pytest-cov==5.0.0
pytest-playwright==0.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2023.3.post1
requests==2.32.5
six==1.17.0
sqlalchemy==2.0.45
typing-extensions==4.15.0
tzdata==2023.3
werkzeug==3.1.4
whitenoise>=6.11.0
sentry-sdk[flask]==1.39.1
Flask-Limiter==3.5.0
Flask-Caching==2.3.1
celery[redis]>=5.3.0
redis>=3.0
aiohttp==3.11.13
aiofiles==24.1.0
pytest-asyncio==0.23.5
